rest/kube/requirements.txt
setuptools
cockroachdb
psycopg2-binary
gunicorn
flask
flask_httpauth
pyyaml
